pyproject.toml
[project]
name = "llm_app_test"
version = "0.2.0b1"
description = "A behavioral testing library for LLM applications that allows developers to write natural language specifications for unit and integration tests. Validate LLM application behavior using plain English assertions in a simple assert(str, str) form factor."
requires-python = ">=3.10"
readme = "README.md"
authors = [
    {name = "Morgan Lee", email = "morganj.lee01@gmail.com"},
]
dependencies = [
    "langchain>=0.3.0",
    "langchain-openai>=0.2.0",
    "langchain-anthropic>=0.2.0",
    "python-dotenv>=1.0.0",
    "pytest>=7.0.0"
]
license = {text = "MIT"}
keywords = ["testing", "llm", "behavioral", "pytest"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Testing"
]

[project.urls]
Homepage = "https://github.com/Shredmetal/llmtest"
Documentation = "https://shredmetal.github.io/llmtest/"
Repository = "https://github.com/Shredmetal/llmtest.git"

[project.optional-dependencies]
dev = [
    "pytest-cov",
    "black",
    "flake8",
    "mkdocs>=1.5.0",
    "mkdocs-material>=9.0.0",
    "build",
    "twine"
]

[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[project.entry-points.pytest11]
llmtest = "llm_app_test.pytest_plugin.plugin"